+++ src/sll/include/sll/_internal/scheduler.h
@@ -15,11 +15,21 @@ typedef struct __SCHEDULER_CPU_DATA{
 	queue_length_t queue_idx;
 	queue_length_t queue_len;
 	sll_internal_thread_index_t tid;
-	sll_event_handle_t sig;
+	sll_event_handle_t evt;
+	sll_lock_handle_t lck;
 	sll_cpu_t id;
 	sll_bool_t wait;
 	sll_thread_index_t queue[];
 } scheduler_cpu_data_t;
+
+
+
+typedef struct __LOAD_BALANCER{
+	scheduler_cpu_data_t** dt;
+	sll_lock_handle_t lck;
+	sll_cpu_t len;
+	sll_cpu_t brk;
+} load_balancer_t;
 
 
 
@@ -32,7 +42,7 @@ void _scheduler_queue_next(void);
 
 
 
-sll_thread_index_t _scheduler_queue_pop(void);
+sll_thread_index_t _scheduler_queue_pop(sll_bool_t lck);
 
 
 
@@ -40,7 +50,7 @@ sll_return_code_t _scheduler_run(void);
 
 
 
-void _scheduler_queue_thread(sll_bool_t any,sll_thread_index_t t);
+void _scheduler_queue_thread(sll_thread_index_t t);